pyproject.toml
[project]
name = "ps-proto"
version = "0.1.0"
description = "OpenCV frame counter with C++ hardware timer integration"
requires-python = ">=3.11"
dependencies = [
    "PySide6==6.8.0.2",
    "opencv-python>=4.12.0.88",
    "pybind11>=3.0.1",
    "setuptools>=80.9.0",
    "numpy>=2.0.0",
    "pywayland>=0.4.18",
    "PyOpenGL>=3.1.0",
]

[dependency-groups]
dev = [
    "pybind11>=3.0.1",
    "types-setuptools>=80.9.0.20250822",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[build-system]
requires = ["setuptools", "pybind11"]
build-backend = "setuptools.build_meta"
